mcp_server_python/requirements.txt
fastapi==0.103.1
uvicorn[standard]==0.23.2
sqlalchemy==2.0.21
alembic==1.12.0
pydantic==2.4.2
pydantic-settings==2.0.3
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
psycopg2-binary==2.9.7
asyncpg==0.28.0
fastapi-pagination==0.12.9
python-multipart==0.0.6
httpx==0.24.1
redis==5.0.0
openai==0.28.1
gitpython==3.1.37
azure-devops==7.1.0b3
structlog==23.1.0
rich==13.5.2
pytest==7.4.2
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.9.1
isort==5.12.0
flake8==6.1.0
mypy==1.5.1